e_transfer_collect.hbs
I made the css for this page. and added the sidenav to the left side.. i think lol. I couldnt see if it worked. so just run it on the server and i can do a quick fix before you present.

I wasnt able to access any of the web pages after logging in.